+++ system-parachains/people/people-kusama/src/weights/mod.rs
@@ -23,6 +23,7 @@ pub mod frame_system;
 pub mod pallet_balances;
 pub mod pallet_collator_selection;
 pub mod pallet_identity;
+pub mod pallet_identity_ops;
 pub mod pallet_message_queue;
 pub mod pallet_multisig;
 pub mod pallet_proxy;
@@ -31,7 +32,6 @@ pub mod pallet_timestamp;
 pub mod pallet_utility;
 pub mod pallet_xcm;
 pub mod paritydb_weights;
-pub mod polkadot_runtime_common_identity_migrator;
 pub mod rocksdb_weights;
 pub mod xcm;